set colors
Major icon colors for wrapper style 2 were set to match bachground.
Style 2 and 3 wrapper alts were set properly to darker shades of
main colors. The header color was updated to match side bar on index.html.
The footer color, that depends on wrapper alt color on the index.html
page and the style1 alt color in the other pages, are set both to
same color.