refix teaching icon
NOTE: Font-Awesome free icons are not rendering, even though all
settings are coirrect. Maybe this is fixable, but might be a headache.
As local images will always render, maybe a good idea to abondon
direct rendering from Font-Awesome as a whole and just add everything
as svg images by hand.